src/server/api/pyproject.toml
[project]
name = "acontext_server"
version = "0.1.0"
description = "Server for acontext -- agent experience layer"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "clickhouse-connect>=0.8.18",
    "elasticsearch>=9.0.2",
    "fastapi[standard]>=0.116.1",
    "openai>=1.95.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "redis>=6.2.0",
    "sqlalchemy[asyncio]>=2.0.41",
    "tiktoken>=0.9.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.2.1",
]
